requirements.txt
# Core dependencies for Resume Tailor app
openai==2.7.1
python-dotenv==1.2.1
python-docx==1.2.0
PySide6==6.10.0
icecream==2.1.4
